fix: more example code compilation errors
- basic_gates.cpp: std::vector<int> -> std::vector<size_t>
- grover_search.cpp: std::vector<int> -> std::vector<size_t>
- grover_search.cpp: add missing search_reg variable

+++ examples/basic_gates.cpp
@@ -174,7 +174,7 @@ void example_qram_with_gates() {
     qram_qutrit::QRAMCircuit qram(addr_size, data_size);
     
     // Initialize memory: [1, 2, 3, 0]
-    std::vector<int> memory = {1, 2, 3, 0};
+    std::vector<size_t> memory = {1, 2, 3, 0};
     qram.set_memory(memory);
     
     std::cout << "QRAM memory:" << std::endl;
